Move security store inside security module + readme
