+++ src/app/core/store/actions/index.ts
@@ -1,5 +1,4 @@
 export * from './busy-indicator.actions';
 export * from './error.actions';
 export * from './cancel.actions';
-export * from './security.actions';
 export * from './snow.actions';
