+++ src/app/core/store/core.selectors.spec.ts
@@ -1,7 +1,5 @@
-import { currentUserSelector, featuresRightsSelector, securityStateSelector, snowStateSelector } from '@app/core/store/core.selectors';
-
-import { busyStateSelector, coreFeatureKey, coreStateSelector, errorStateSelector } from './core.selectors';
-import { BusyState, CoreState, ErrorState, SecurityState } from './models';
+import { busyStateSelector, coreFeatureKey, coreStateSelector, errorStateSelector, snowStateSelector } from './core.selectors';
+import { BusyState, CoreState, ErrorState, } from './models';
 import { SnowState } from './models/snow.state';
 
 describe('core selector', () => {
@@ -11,7 +9,6 @@ describe('core selector', () => {
       busy: { actionsInProgress: [] },
       error: { lastError: 'test ' },
       snow: undefined,
-      security: undefined
     };
 
     const extractState = coreStateSelector({
@@ -60,59 +57,4 @@ describe('core selector', () => {
 
     expect(result).toEqual(snowState);
   });
-
-  it('should extract security state', () => {
-    const securityState = {
-      currentUser: { name: 'name', rights: ['r'] },
-      featuresConfiguration: [
-        { name: 'f', rights: ['r'] }
-      ],
-      noRightPath: ''
-    };
-
-    const result = securityStateSelector({
-      [coreFeatureKey]: {
-        security: securityState
-      }
-    });
-
-    expect(result).toEqual(securityState);
-  });
-
-  it('should extract current user state', () => {
-    const securityState = {
-      currentUser: { name: 'name', rights: ['r'] },
-      featuresConfiguration: [
-        { name: 'f', rights: ['r'] }
-      ],
-      noRightPath: ''
-    };
-
-    const result = currentUserSelector({
-      [coreFeatureKey]: {
-        security: securityState
-      }
-    });
-
-    expect(result).toEqual({ name: 'name', rights: ['r'] });
-  });
-
-  it('should extract features rights state', () => {
-    const securityState = {
-      currentUser: { name: 'name', rights: ['r'] },
-      featuresConfiguration: [
-        { name: 'f', rights: ['r'] }
-      ],
-      noRightPath: ''
-    };
-
-    const result = featuresRightsSelector({
-      [coreFeatureKey]: {
-        security: securityState
-      }
-    });
-
-    expect(result).toEqual([{ name: 'f', rights: ['r'] }]);
-  });
-
 });
