+++ src/app/core/store/core.selectors.ts
@@ -9,7 +9,4 @@ export const coreStateSelector = createFeatureSelector<CoreState>(coreFeatureKey
 export const busyStateSelector = createSelector(coreStateSelector, (state: CoreState) => state.busy);
 export const errorStateSelector = createSelector(coreStateSelector, (state: CoreState) => state.error);
 export const snowStateSelector = createSelector(coreStateSelector, (state: CoreState) => state.snow);
-export const securityStateSelector = createSelector(coreStateSelector, (state: CoreState) => state.security);
 
-export const currentUserSelector = createSelector(securityStateSelector, (state) => state.currentUser);
-export const featuresRightsSelector = createSelector(securityStateSelector, (state) => state.featuresConfiguration);
