+++ src/app/core/store/models/core.state.ts
@@ -1,11 +1,9 @@
 import { BusyState } from './busy.state';
 import { ErrorState } from './error.state';
-import { SecurityState } from './security.state';
 import { SnowState } from './snow.state';
 
 export interface CoreState {
   busy: BusyState;
   error: ErrorState;
   snow: SnowState;
-  security: SecurityState;
 }
